feat: add getAvailableVCStores RPC method
