+++ snap.manifest.json
@@ -7,7 +7,7 @@
     "url": "https://github.com/blockchain-lab-um/ssi-snap.git"
   },
   "source": {
-    "shasum": "ZJDE2fPOcPYuzrwZ4HQNsvm34Ir0RRgMhDyOkCiGcOU=",
+    "shasum": "SJ4qBkHk2jru3XiXYNjjx88/MW4//ZunRda39WyUD+s=",
     "location": {
       "npm": {
         "filePath": "dist/snap.js",
