+++ src/index.ts
@@ -1,8 +1,8 @@
 import { OnRpcRequestHandler } from '@metamask/snap-types';
-import { togglePopups, changeInfuraToken } from './rpc/configure';
-import { getVCs } from './rpc/getVCs';
-import { getVP } from './rpc/getVP';
-import { saveVC } from './rpc/saveVC';
+import { togglePopups, changeInfuraToken } from './rpc/snap/configure';
+import { getVCs } from './rpc/vc/getVCs';
+import { getVP } from './rpc/vc/getVP';
+import { saveVC } from './rpc/vc/saveVC';
 import {
   isValidChangeInfuraTokenRequest,
   isValidGetVCsRequest,
@@ -10,15 +10,15 @@ import {
   isValidSaveVCRequest,
   isValidSwitchMethodRequest,
 } from './utils/params';
-import { switchMethod } from './rpc/switchMethod';
-import { init } from './rpc/init';
-import { getDid } from './rpc/getDID';
-import { getMethod } from './rpc/getMethod';
-import { getAvailableMethods } from './rpc/getAvailableMethods';
-import { getVCStore } from './rpc/getVCStore';
-import { setVCStore } from './rpc/setVCStore';
+import { switchMethod } from './rpc/did/switchMethod';
+import { init } from './rpc/snap/init';
+import { getDid } from './rpc/did/getDID';
+import { getMethod } from './rpc/did/getMethod';
+import { getAvailableMethods } from './rpc/did/getAvailableMethods';
+import { getVCStore } from './rpc/vcStore/getVCStore';
+import { setVCStore } from './rpc/vcStore/setVCStore';
 import { clear } from './veramo/plugins/ceramicDataStore/ceramicDataStore';
-import { DID } from 'dids';
+import { getAvailableVCStores } from './rpc/vcStore/getAvailableVCStores';
 
 export const onRpcRequest: OnRpcRequestHandler = async ({
   origin,
@@ -67,6 +67,8 @@ export const onRpcRequest: OnRpcRequestHandler = async ({
       return await getVCStore();
     case 'setVCStore':
       return await setVCStore();
+    case 'getAvailableVCStores':
+      return getAvailableVCStores();
     default:
       throw new Error('Method not found.');
   }
