+++ src/rpc/did/getDID.ts
@@ -1,4 +1,4 @@
-import { getCurrentDid } from '../utils/didUtils';
+import { getCurrentDid } from '../../utils/didUtils';
 
 export async function getDid(): Promise<string> {
   return await getCurrentDid();
