+++ src/rpc/did/getMethod.ts
@@ -1,4 +1,4 @@
-import { getCurrentMethod } from '../utils/didUtils';
+import { getCurrentMethod } from '../../utils/didUtils';
 
 export async function getMethod(): Promise<string> {
   return await getCurrentMethod();
