+++ src/rpc/vcStore/getVCStore.ts
@@ -1,0 +1,5 @@
+import { getCurrentVCStore } from '../../utils/didUtils';
+
+export async function getVCStore(): Promise<string> {
+  return await getCurrentVCStore();
+}
