--- src/veramo/plugins/availableDataStores.ts
@@ -1,1 +1,0 @@
-export const availableDataStores = ['snap', 'ceramic'] as const;
