+++ src/veramo/plugins/availableVCStores.ts
@@ -1,0 +1,1 @@
+export const availableVCStores = ['snap', 'ceramic'] as const;
